#78 More Refactoring
Committing this because Its actually functional so if I break something more, I can come back to this commit
