+++ .github/pull_request_template.md
@@ -1,12 +1,12 @@
-##Changes:
+## Changes:
 
 
 
-#Notes:
+### Notes:
 
 
 
-#Tests/ScreenShots
+### Tests/ScreenShots
